ORM Update
I updated some of the ORM code so that it can match the data type of our primary keys (int -> varchar(255).
Make sure you merge with my branch before you attempt to use the orms otherwise u will get the same errors as me.
i updated the code for all 8 of the tables.
bare in mind that i might have missed some lines so if u experience problems. go to the vb file for the specific table. all the primary and foreign keys should be string and when you scroll you will see parameters with numbers. for all primary keys it should be like this format for example
"@Child_ID", 22, 255, "Child_ID". 22 represents varchar and the 255 makes it varchar(255)
but you wont have to worry about this if i indeed updated everything